fix(react-native): return update fetch promises directly
