+++ .changeset/direct-update-fetch.md
@@ -1,0 +1,5 @@
+---
+"@hot-updater/react-native": patch
+---
+
+Return update check fetch promises directly instead of wrapping them in task pipelines.
