+++ packages/react-native/src/fetchUpdateInfo.ts
@@ -1,7 +1,5 @@
 import type { AppUpdateInfo } from "@hot-updater/core";
 
-import { task, withRetry, withTimeout } from "./utils/task";
-
 const requireResponse = (response: Response | undefined) => {
   if (!response) {
     throw new Error("Fetch returned no response");
@@ -16,7 +14,7 @@ const parseUpdateInfo = (response: Response): Promise<AppUpdateInfo | null> => {
   return response.json();
 };
 
-export const fetchUpdateInfo = async ({
+export const fetchUpdateInfo = ({
   url,
   requestHeaders,
   requestTimeout = 5000,
@@ -29,22 +27,24 @@ export const fetchUpdateInfo = async ({
     "Content-Type": "application/json",
     ...requestHeaders,
   };
+  const controller = new AbortController();
+  const timeoutId = setTimeout(() => {
+    controller.abort();
+  }, requestTimeout);
 
-  const getUpdateResponse = task<Response | undefined>(
-    (signal?: AbortSignal) => {
-      return fetch(url, {
-        signal,
-        headers,
-      });
-    },
-  );
-
-  const response = await getUpdateResponse
-    .pipe(
-      withTimeout(requestTimeout),
-      withRetry(1, (response) => !response),
-    )
-    .run();
-
-  return parseUpdateInfo(requireResponse(response));
+  return fetch(url, {
+    signal: controller.signal,
+    headers,
+  })
+    .then(requireResponse)
+    .then(parseUpdateInfo)
+    .catch((error: unknown) => {
+      if (error instanceof Error && error.name === "AbortError") {
+        throw new Error("Request timed out");
+      }
+      throw error;
+    })
+    .finally(() => {
+      clearTimeout(timeoutId);
+    });
 };
